requirements.txt
Flask
Pillow
numpy
werkzeug
